feat: final review Changes done
The <ul> tag (breadcrumb) in restaurant.html needs an appropriate aria label.
In index.html, the title of the application and the restaurant's name have different levels of importance, so they should have different levels of heading style. E.g. h1 for the application title and h2 for the restaurant's name.
In restaurant.html, the title of the application, the restaurant's name and the title of reviews have different levels of importance, so they should have different levels of heading style. E.g. h1 for the application title, h2 for the restaurant's name and h3 for the title of reviews.
Making the alt value descriptive enough stating it is an image of the restaurant. This is very important for people using screen readers.